Incorporate destroy-circular logic. Fixes #2.
For serialized AssertionErrors, it is possible that the `actual`
and `expected` properties contain circular references, causing
a thrown Error when you serialize the result with JSON.stringify().

This destroys circular references on the copied Object. Replacing
them with a string (`[Circular]`).

Also, `JSON.stringify(function() {})` returns `undefined`, and
will drop any functions that are attached to an object it is serializing.
This mimics the latter behavior, but if a function is passed in as the top
level argument, it returns a string (`[Function: fnName]`). Since
the most common use for this is to transmit thrown values for logging,
it makes sense to provide this helpful String instead of `undefined`.
(All this despite the fact it is a terrible idea to throw a function).

Based on:
  https://github.com/jonpacker/destroy-circular/pull/1

Fixes: #2